move type
